+++ src/handlers/mod.rs
@@ -2,7 +2,6 @@ pub mod cleanup_position_sl_tp;
 pub mod cleanup_position_stop_loss;
 pub mod cleanup_position_take_profit;
 pub mod create_ixs;
-pub mod grpf;
 pub mod liquidate_long;
 pub mod liquidate_short;
 pub mod sl_long;
@@ -12,6 +11,6 @@ pub mod tp_short;
 
 pub use {
     cleanup_position_sl_tp::*, cleanup_position_stop_loss::*, cleanup_position_take_profit::*,
-    create_ixs::*, grpf::*, liquidate_long::liquidate_long, sl_long::sl_long, sl_short::sl_short,
+    create_ixs::*, liquidate_long::liquidate_long, sl_long::sl_long, sl_short::sl_short,
     tp_long::tp_long, tp_short::tp_short,
 };
